backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0

# LLM SDKs
openai>=1.35.0
anthropic>=0.30.0
google-genai>=0.8.0
cohere>=5.5.0
mistralai>=1.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Security & Crypto
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0

# HTTP
httpx>=0.27.0

# Redis (Phase 2 prep)
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
langsmith>=0.1.0
langchain>=0.2.0
